Users dev

+++ components/common/table/TableTopicsComponent.vue
@@ -21,13 +21,17 @@
           </div>
           <div class="d-flex me-auto cursor-pointer ms-3 text-danger"
             v-if="actionDelete && useCurrentsRole(currentRole,[ROLES.ROLE_ADMIN, ROLES.ROLE_TOPIC_DELETE])"
-            @click="deleteTopic(item.id)">
+            data-bs-toggle="modal" :data-bs-target="'#' + modalSelectorId" @click="setItem(item)">
             <delete-icon />
             <span class="ms-1">Xóa</span>
           </div>
         </div>
       </div>
     </div>
+    <ConfirmDelete :idModal="modalSelectorId" :objectName="topicConfirm.name" :objectType="objectType"
+      :btnTitleConfirm="btnTitleConfirm" :issueText="issueText"
+      :idObject="topicConfirm.id" @func-delete="deleteTopic(topicConfirm.id)"
+    />
   </div>
 </template>
 <script>
@@ -36,43 +40,66 @@ import DeleteIcon from "~~/assets/images/icons/actions/DeleteIcon.vue";
 import { useCurrentsRole } from "~~/services/common.js"
 import { ROLES } from "~~/constants/roles.js";
 import TopicService from "~~/services/model/topic.service";
+import ConfirmDelete from '~~/components/common/modal/ConfirmDelete.vue';
 
 export default {
   components: {
     EditIcon,
     DeleteIcon,
+    ConfirmDelete
   },
   props: ["headers", "items", "actionEdit", "actionDelete", "page", "size"],
+  data() {
+    return {
+      modalSelectorId: "topicConfirmDelete",
+      objectType: "chuyên mục",
+      issueText: "Các chuyên mục có bài viết sẽ không thể xóa",
+      btnTitleConfirm: "Tôi hiểu hậu quả, xóa chuyên mục này"
+    }
+  },
   setup() {
     const currentRole = useCurrentRole();
     const { $showToast } = useNuxtApp();
-
+    const topicConfirm = reactive({
+      id: "",
+      name: ""
+    });
+    function setItem(topic) {
+      if (topic) {
+        topicConfirm.id = topic.id;
+        topicConfirm.name = topic.name;
+      }
+    }
     function deleteTopic(topicId) {
-      TopicService.deleteById(topicId)
-        .then(response => {
-          let data = response.data;
-          if (data) {
-            if (data.code == 200) {
-              $showToast("Xóa chuyên mục thành công!", "success", 3000);
-              location.reload();
-            } else {
-              
-              $showToast("Xóa chuyên mục không thành công!", "error", 3000);
+      if(topicId) {
+        TopicService.deleteById(topicId)
+          .then(response => {
+            let data = response.data;
+            if (data) {
+              if (data.code == 200) {
+                $showToast("Xóa chuyên mục thành công!", "success", 3000);
+                location.reload();
+              } else {
+                
+                $showToast("Xóa chuyên mục không thành công!", "error", 3000);
+              }
             }
-          }
-        })
-        .catch(error => {
-          $showToast("Xóa chuyên mục không thành công!", "error", 3000);
-          console.log(error.toString());
-        })
-        ;
+          })
+          .catch(error => {
+            $showToast("Xóa chuyên mục không thành công!", "error", 3000);
+            console.log(error.toString());
+          })
+          ;
+      }
     }
     return {
+      topicConfirm,
       currentRole,
       ROLES,
 
       deleteTopic,
       useCurrentsRole,
+      setItem
     }
   },
 };
